move the Dummy-Model to the package base;
