+++ tests/__init__.py
@@ -1,0 +1,28 @@
+# Copyright (c) individual contributors.
+# All rights reserved.
+#
+# This is free software; you can redistribute it and/or modify it
+# under the terms of the GNU Lesser General Public License as
+# published by the Free Software Foundation; either version 3 of
+# the License, or any later version.
+#
+# This software is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+# Lesser General Public License for more details. A copy of the
+# GNU Lesser General Public License is distributed along with this
+# software and can be found at http://www.gnu.org/licenses/lgpl.html
+from zipfile import ZipFile
+from io import BytesIO
+
+
+class Dummy:
+    def toBytes(self):
+        data = BytesIO()
+        f = ZipFile(data, "w")
+        model_file = f.open("__model__.py", "w")
+        model_file.write(b"")
+        model_file.close()
+        f.close()
+
+        return data.getvalue()
